This release is v0.2 It includes the G4Sipm module but not used in the simulation code.
MDM_DetectionConctruction.cc
It updates the contruction file to include the latest understanding of photon reflection configuration. Material needs to be assgined a RINDEX property to allow the photon entry.

MDM_PHysicsList.cc
Completly replace the original code with the code from example OpNovice.
The code has much in-depth control of the physics process and particle construction